% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{apply_initial_threshold}
\alias{apply_initial_threshold}
\title{Apply threshold}
\usage{
apply_initial_threshold(eSVD_obj, pval_thres = 0.01, verbose = 0)
}
\arguments{
\item{eSVD_obj}{\code{eSVD} object outputed from \code{initialize_esvd}.}

\item{pval_thres}{Numeric between 0 and 1 (inclusive).}

\item{verbose}{Integer}
}
\value{
\code{eSVD} object with added elements \code{fit_Init} and
\code{latest_Fit}
}
\description{
Based on a p-value threshold, determine which sets of coefficients to
use for each gene. Then, based on the residuals,
initialize the latent embedding according to the SVD.
}
