% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{compute_posterior}
\alias{compute_posterior}
\title{Compute posterior according to Gamma-Poisson model}
\usage{
compute_posterior(input_obj, ...)
}
\arguments{
\item{input_obj}{Main object}

\item{...}{Additional parameters}
}
\value{
Output dependent on class of \code{input_obj}
}
\description{
Generic function interface
}
