% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_test_statistic.R
\name{compute_test_statistic}
\alias{compute_test_statistic}
\title{Compute test statistics}
\usage{
compute_test_statistic(input_obj, ...)
}
\arguments{
\item{input_obj}{Main object}

\item{...}{Additional parameters}
}
\value{
Output dependent on class of \code{input_obj}
}
\description{
Generic function interface
}
