% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pvalue.R
\name{.compute_df}
\alias{.compute_df}
\title{Compute the degree of freedom}
\usage{
.compute_df(input_obj)
}
\arguments{
\item{input_obj}{\code{eSVD} object outputed from \code{compute_test_statistic}.}
}
\value{
a named vector of degree-of-freedom values, one for each gene
}
\description{
This is an intermediary function used in \code{compute_pvalue}
}
