% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.compute_principal_angle}
\alias{.compute_principal_angle}
\title{Compute principal angle}
\usage{
.compute_principal_angle(tmp_mat)
}
\arguments{
\item{tmp_mat}{a matrix with \code{n} rows (for \code{n} samples) and \code{2} columns,
where the first column represents the observed data and the second column represents its
corresponding predicted values}
}
\value{
numeric
}
\description{
Compute principal angle
}
