% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{.determine_initial_matrix}
\alias{.determine_initial_matrix}
\title{Initialize the matrix of natural parameters}
\usage{
.determine_initial_matrix(
  dat,
  k,
  family,
  nuisance_param_vec,
  max_val = NA,
  tol = 0.001
)
}
\arguments{
\item{dat}{dataset where the \code{n} rows represent cells and \code{d} columns represent genes.}

\item{k}{positive integer less than \code{min(c(nrow(dat), ncol(dat)))}}

\item{family}{character (\code{"gaussian"}, \code{"exponential"}, \code{"poisson"}, \code{"neg_binom"},
or \code{"curved gaussian"})}

\item{nuisance_param_vec}{length-\eqn{p}
vector of numerics representing nuisance parameters (for \code{family = "neg_binom"} or
\code{family = "curved_gausian"})}

\item{max_val}{maximum magnitude of the inner product}

\item{tol}{numeric
or \code{"curved gaussian"} for \code{family}}
}
\value{
\code{n} by \code{p} matrix
}
\description{
This function first transforms each entry in \code{dat} according to the inverse function that maps
natural parameters to their expectation (according to \code{eSVD:::.mean_transformation}) and then
uses \code{eSVD:::.project_rank_feasibility} to get a rank-\code{k} approximation of this matrix
that lies within the domain of \code{family}
}
