% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{.gradient_mat}
\alias{.gradient_mat}
\title{Gradient of the objective function}
\usage{
.gradient_mat(dat, nat_mat, ...)
}
\arguments{
\item{dat}{dataset where the \code{n} rows represent cells and \code{d} columns represent genes}

\item{nat_mat}{\code{n} by \code{d} matrix where each entry represents the natural parameter of the corresponding entry in \code{dat}}

\item{...}{other parameters}
}
\value{
\code{n} by \code{d} matrix
}
\description{
Computes the gradient for a particular model (based on the
class of \code{dat}) of the \code{.evaluate_objective_mat} function.
}
\details{
Note, \code{dat} is NOT allowed to have any \code{NA} values for this
function.
}
