% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{.matrix_completion}
\alias{.matrix_completion}
\title{Fill in missing values}
\usage{
.matrix_completion(dat, k)
}
\arguments{
\item{dat}{dataset where the \code{n} rows represent cells and \code{d} columns represent genes}

\item{k}{positive integer less than \code{min(c(nrow(dat), ncol(dat)))}}
}
\value{
a \code{n} by \code{p} matrix
}
\description{
Uses \code{softImpute::softImpute} to fill in all the possible missing values.
This function enforces all the resulting entries to be non-negative.
}
