% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparameterization.R
\name{.reparameterize}
\alias{.reparameterize}
\title{Function to reparameterize two matrices}
\usage{
.reparameterize(x_mat, y_mat, equal_covariance)
}
\arguments{
\item{x_mat}{matrix of dimension \code{n} by \code{k}}

\item{y_mat}{matrix of dimension \code{p} by \code{k}}

\item{equal_covariance}{boolean}
}
\value{
list of two matrices
}
\description{
test
}
\details{
Designed to output matrices of the same dimension as \code{x_mat}
and \code{y_mat}, but linearly transformed so \code{x_mat \%*\% t(y_mat)}
is preserved but either \code{x_mat \%*\% t(x_mat)} is diagonal and equal to
\code{y_mat \%*\% t(y_mat)} (if \code{equal_covariance} is \code{FALSE})
or \code{x_mat \%*\% t(x_mat)/nrow(x_mat)} is diagonal and equal to
\code{y_mat \%*\% t(y_mat)/nrow(y_mat)} (if \code{equal_covariance} is \code{TRUE})
}
