% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance.R
\name{estimate_nuisance}
\alias{estimate_nuisance}
\title{Estimate nuisance values}
\usage{
estimate_nuisance(input_obj, ...)
}
\arguments{
\item{input_obj}{Main object}

\item{...}{Additional parameters}
}
\value{
Output dependent on class of \code{input_obj}
}
\description{
Generic function interface
}
