% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esvd-package.R
\docType{package}
\name{eSVD2-package}
\alias{eSVD2-package}
\title{eSVD2 package, C++ settings}
\description{
Temporary description

Maintainer: None
}
\author{
Me \email{kevinl1@wharton.upenn.edu}
}
\keyword{package}
