% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_test.R
\name{fisher_test}
\alias{fisher_test}
\title{Fisher's exact test}
\usage{
fisher_test(set1_genes, set2_genes, all_genes, verbose = 0)
}
\arguments{
\item{set1_genes}{A vector of characters of your resulting gene list}

\item{set2_genes}{A vector of characters of a gene list you are comparing against}

\item{all_genes}{A vector of characters of all the genes in your "universe". This should include all the genes in \code{set1_genes} and \code{set2_genes}.}

\item{verbose}{Numeric}
}
\value{
A list of: \code{pvalue} for the p-value,
\code{set_bg_len} for the number of genes in \code{all_genes} but not in \code{set2_genes},
\code{set_overlap_len} for the number of genes in both \code{set1_genes} and \code{set2_genes},
\code{set1_len} for number of genes in \code{set1_genes}, and
\code{set2_len} for number of genes in \code{set2_genes}.
}
\description{
Perform a Fisher's exact test. This formula is from https://www.pathwaycommons.org/guide/primers/statistics/fishers_exact_test/
}
\examples{
all_genes <- paste0("gene", 1:1000)
set1_genes <- paste0("gene", 1:50)
set2_genes <- paste0("gene", c(1:20, 501:530))
res <- fisher_test(set1_genes = set1_genes,
                   set2_genes = set2_genes,
                   all_genes = all_genes)
res$pvalue
res$set_overlap_len
}
