% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gandal_df}
\alias{gandal_df}
\title{Gandal et al. results}
\format{
## `gandal_df`
A data frame with 24836 rows (genes) and 3 columns:
\describe{
  \item{external_gene_name}{Gene symbol}
  \item{WholeCortex_ASD_logFC}{Estimated log fold change}
  \item{WholeCortex_ASD_FDR}{FDR q-value}
}
}
\source{
<https://www.nature.com/articles/s41586-022-05377-7>
}
\usage{
gandal_df
}
\description{
From https://www.nature.com/articles/s41586-022-05377-7,
Supplementary Data 3 (i.e., \code{41586_2022_5377_MOESM5_ESM.xlsx})
}
\examples{
\dontrun{
# How the data was loaded
df <- openxlsx::read.xlsx("41586_2022_5377_MOESM5_ESM.xlsx",
                          sheet = "DEGene_Statistics")
gandal_df <- df[,c("external_gene_name", "WholeCortex_ASD_logFC", "WholeCortex_ASD_FDR")]
usethis::use_data(gandal_df)
}
}
\keyword{datasets}
