% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{housekeeping_df}
\alias{housekeeping_df}
\title{Hounkpe et al. housekeeping genes}
\format{
## `housekeeping_df`
A data frame with 1003 rows (genes) and 1 column:
\describe{
  \item{Gene}{Gene name}
}
}
\source{
<https://www.science.org/doi/10.1126/science.aav8130>
}
\usage{
housekeeping_df
}
\description{
From https://academic.oup.com/nar/article/49/D1/D947/5871367#supplementary-data,
gkaa609_Supplemental_Files, specifically, \code{Supplementary_Table1.xlsx})
}
\examples{
\dontrun{
# How the data was loaded
df <- openxlsx::read.xlsx("Supplementary_Table1.xlsx",
sheet = "Gene Model")
colnames(df) <- df[1,1]
housekeeping_df <- df[-1,,drop = FALSE]
usethis::use_data(housekeeping_df)
}
}
\keyword{datasets}
