% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initalization_default}
\alias{initalization_default}
\title{Initialization defaults}
\usage{
initalization_default(method = "kmean_rows", max_val = NA, tol = 0.001)
}
\arguments{
\item{method}{character (\code{"kmean_rows"})}

\item{max_val}{maximum magnitude of the inner product (positive numeric). This parameter
could be \code{NA}.}

\item{tol}{small positive value, which also controls (amongst other numerical things)
the smallest magnitude of the inner product}
}
\value{
a list of values, of class \code{initialization_param}
}
\description{
Initialization defaults
}
