% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initialization_param}
\alias{initialization_param}
\title{Initialization defaults}
\usage{
initialization_param(
  init_method = "kmean_rows",
  library_size_method = "total_read",
  nuisance_est_method = "mom",
  max_val = NA,
  tol = 0.001
)
}
\arguments{
\item{max_val}{maximum magnitude of the inner product (positive numeric). This parameter
could be \code{NA}.}

\item{tol}{small positive value, which also controls (amongst other numerical things)
the smallest magnitude of the inner product}

\item{method}{character (\code{"kmean_rows"})}
}
\value{
a list of values, of class \code{initialization_param}
}
\description{
Initialization defaults
}
