% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initialize_esvd}
\alias{initialize_esvd}
\title{Initialize eSVD}
\usage{
initialize_esvd(
  dat,
  covariates,
  metadata_individual,
  bool_intercept = FALSE,
  case_control_variable = NULL,
  k = 30,
  lambda = 0.01,
  library_size_variable = "Log_UMI",
  offset_variables = "Log_UMI",
  metadata_case_control = NULL,
  verbose = 0
)
}
\arguments{
\item{dat}{Dataset (either \code{matrix} or \code{dgCMatrix}) where the \eqn{n} rows represent cells
and \eqn{p} columns represent genes.
The rows and columns of the matrix should be named.}

\item{covariates}{\code{matrix} object with \eqn{n} rows with the same rownames as \code{dat} where the columns
represent the different covariates.
Notably, this should contain only numerical columns (i.e., all categorical
variables should have already been split into numerous indicator variables), and all the columns
in \code{covariates} will (strictly speaking) be included in the eSVD matrix factorization model.}

\item{metadata_individual}{\code{factor} vector of length \eqn{n} that denotes which cell originates from which individual.}

\item{bool_intercept}{Boolean on whether or not an intercept will be included as a covariate.}

\item{case_control_variable}{A string of the column name of \code{covariates} which depicts the case-control
status of each cell. Notably, this should be a binary variable where a \code{1}
is hard-coded to describe case, and a \code{0} to describe control.}

\item{k}{Number of latent dimensions.}

\item{lambda}{Ridge penalty used by \code{glmnet::glmnet} on every covariate that is
neither the intercept nor an offset; must lie in \eqn{[10^{-4}, 10^4]}.}

\item{library_size_variable}{A string of the variable name (which must be in \code{covariates}) of which variable denotes the sequenced (i.e., observed) library size.}

\item{offset_variables}{A vector of strings depicting which column names in \code{covariate} will
be set to have a coefficient of \code{1} automatically (i.e., there will be no estimation
of their coefficient).}

\item{metadata_case_control}{(Optional) vector of length \eqn{n} with values strictly 0 or 1 that denotes if a cell is from cases or controls.
By default, this is set to \code{NULL} since the code will extract this information from \code{covariates}.}

\item{verbose}{Integer}
}
\value{
\code{eSVD} object with elements \code{dat}, \code{covariates},
\code{param}, \code{fit_Init} (an \code{eSVD_Fit} with \code{x_mat},
\code{y_mat}, \code{z_mat}), \code{latest_Fit} (\code{"fit_Init"}),
\code{case_control} and \code{individual}
}
\description{
For each gene, this function fits one ridge-penalized Poisson GLM of the counts on
\code{covariates} (with \code{offset_variables} as offsets; \code{stats::glm} when
at most one covariate is left to estimate) to obtain \code{z_mat}, and then takes a
rank-\code{k} SVD of \code{log1p(dat)} minus the fitted covariate effects to obtain
\code{x_mat} and \code{y_mat}. The result is stored as \code{fit_Init}.
}
\examples{
set.seed(10)
sim <- generate_null(cell_per_person = 15, num_genes = 40,
                     num_individuals = 8)
esvd_obj <- initialize_esvd(dat = sim$obs_mat,
                            covariates = sim$covariates,
                            metadata_individual = sim$metadata_individual,
                            case_control_variable = "CC",
                            bool_intercept = TRUE,
                            k = 2,
                            lambda = 0.1)
esvd_obj <- reparameterization_esvd_covariates(input_obj = esvd_obj,
                                               fit_name = "fit_Init",
                                               omitted_variables = "Log_UMI")
names(esvd_obj)
dim(esvd_obj$fit_Init$x_mat)
}
