% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_estimation.R
\name{initialize_nuisance_param}
\alias{initialize_nuisance_param}
\title{Estimating the nuisance parameters given an initial estimate of natural matrix}
\usage{
initialize_nuisance_param(dat, init_nat_mat, family, library_size_vec)
}
\arguments{
\item{dat}{dataset where the \eqn{n} rows represent cells and \eqn{p} columns represent genes}

\item{init_nat_mat}{initial estimate of the natural parameter matrix, of also size \eqn{n \times p}}

\item{family}{A character string, one of \code{"gaussian"}, \code{"exponential"},
\code{"poisson"}, \code{"neg_binom"}, \code{"curved_gaussian"},
and \code{"bernoulli"}.}

\item{library_size_vec}{either \code{NA} or a single numeric (default is \code{1}) or
a length-\eqn{n} vector of numerics.
If \code{NA}, the library size will be estimated.}
}
\value{
vector of length \eqn{p}, representing the nuisance parameters, one for each
of the \eqn{p} genes
}
\description{
Estimating the nuisance parameters given an initial estimate of natural matrix
}
