% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{opt_esvd}
\alias{opt_esvd}
\title{Main optimization function for eSVD}
\usage{
opt_esvd(
  x_init,
  y_init,
  dat,
  family,
  method = c("newton", "lbfgs"),
  nuisance_param_vec = NA,
  library_size_vec = 1,
  max_iter = 100,
  tol = 1e-06,
  verbose = 0,
  ...
)
}
\arguments{
\item{x_init}{initial estimate of a \eqn{n \times k} matrix for the \eqn{k}-dimensional
embedding for the \eqn{n} cells}

\item{y_init}{initial estimate of a \eqn{p \times k} matrix for the \eqn{k}-dimensional
embedding for the \eqn{p} genes}

\item{dat}{dataset where the \eqn{n} rows represent cells and \eqn{p} columns represent genes}

\item{family}{A character string, one of \code{"gaussian"}, \code{"exponential"},
\code{"poisson"}, \code{"neg_binom"}, \code{"curved_gaussian"},
and \code{"bernoulli"}.}

\item{method}{A character string indicating the optimization method,
either \code{"newton"} or \code{"lbfgs"}.}

\item{nuisance_param_vec}{either \code{NA} or a single numeric or a length-\eqn{p}
vector of numerics representing nuisance parameters (for \code{family = "neg_binom"} or
\code{family = "curved_gausian"}).}

\item{library_size_vec}{either \code{NA} or a single numeric (default is \code{1}) or
a length-\eqn{n} vector of numerics.
If \code{NA}, the library size will be estimated.}

\item{max_iter}{positive integer}

\item{tol}{small positive number}

\item{verbose}{non-negative integer}

\item{...}{additional parameters, not used}
}
\value{
a list with elements \code{x_mat} and \code{y_mat} (and others), representing the two
latent matrices
}
\description{
Main optimization function for eSVD
}
