% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_helper.R
\name{opt_yz}
\alias{opt_yz}
\title{Optimize Y and Z given X and C}
\usage{
opt_yz(
  YZ_init,
  XC,
  k,
  fixed_cols,
  loader,
  family,
  s,
  gamma,
  l2peny,
  l2penz,
  verbose = 0,
  inplace = FALSE,
  ...
)
}
\arguments{
\item{YZ_init}{Initial value for the `[Y Z]` matrix, Y [p x k], Z [p x r]}

\item{XC}{The `[X C]` matrix, X [n x k], C [n x r]}

\item{k}{Number of columns in X and Y}

\item{fixed_cols}{Which columns in YZ need to be fixed}

\item{loader}{The data loader, typically returned by data_loader()}

\item{family}{A family object, typically returned by esvd_family()}

\item{s}{The library size vector, [n x 1]}

\item{gamma}{The nuisance parameter vector, [p x 1]}

\item{l2peny}{The l2 penalty parameter for Y, a scalar}

\item{l2penz}{The l2 penalty parameter for Z, a scalar}

\item{verbose}{Verbosity parameter}

\item{inplace}{Whether the input YZ_init will be modified and returned}

\item{...}{Additional parameters}
}
\value{
The optimized `[Y Z]` matrix, p by (k + r), with the columns in
`fixed_cols` unchanged. It carries an integer attribute
`num_linesearch_failed`, the number of columns whose Newton line search
found no descent step.
}
\description{
Optimize Y and Z given X and C
}
\keyword{internal}
