% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparameterization.R
\name{reparameterization_esvd_covariates}
\alias{reparameterization_esvd_covariates}
\title{Reparameterize eSVD object}
\usage{
reparameterization_esvd_covariates(
  input_obj,
  fit_name,
  omitted_variables = NULL,
  verbose = 0
)
}
\arguments{
\item{input_obj}{\code{eSVD} object, after either \code{initialize_esvd()} or
\code{opt_esvd()}}

\item{fit_name}{The name of the fit in \code{input_obj} that you wish to reparameterize. This should be in \code{names(input_obj)}}

\item{omitted_variables}{Either \code{NULL} (the default) or variables in \code{input_obj$covariates} that should not be reparameterized}

\item{verbose}{Integer.}
}
\value{
\code{eSVD} object after adjusting the fit in \code{fit_name}.
}
\description{
Reparameterize eSVD object
}
\examples{
set.seed(10)
sim <- generate_null(cell_per_person = 15, num_genes = 40,
                     num_individuals = 8)
esvd_obj <- initialize_esvd(dat = sim$obs_mat,
                            covariates = sim$covariates,
                            metadata_individual = sim$metadata_individual,
                            case_control_variable = "CC",
                            bool_intercept = TRUE,
                            k = 2,
                            lambda = 0.1)
esvd_obj <- reparameterization_esvd_covariates(input_obj = esvd_obj,
                                               fit_name = "fit_Init",
                                               omitted_variables = "Log_UMI")
# x_mat is now orthogonal to the retained covariates
max(abs(crossprod(esvd_obj$fit_Init$x_mat,
                  esvd_obj$covariates[, "Sex"])))
}
