% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sfari_df}
\alias{sfari_df}
\title{SFARI genes}
\format{
## `sfari_df`
A data frame with 1176 rows (genes) and 10 columns:
\describe{
  \item{status}{See website for details}
  \item{gene.symbol}{Gene symbol}
  \item{gene.name}{Full gene name}
  \item{ensembl.id}{Gene ID}
  \item{chromosome}{Chromosome the gene is on}
  \item{genetic.category}{See website for details}
  \item{gene.score}{See website for details}
  \item{syndromic}{See website for details}
  \item{eagle}{See website for details}
  \item{number.of.reports}{See website for details}
}
}
\source{
<https://gene.sfari.org/database/human-gene/>
}
\usage{
sfari_df
}
\description{
From https://gene.sfari.org/database/human-gene/, after clicking on "Download this dataset",
accessed on June 15, 2024
}
\examples{
\dontrun{
# How the data was loaded
sfari_df <- read.csv("SFARI-Gene_genes_03-28-2024release_06-16-2024export.csv")
usethis::use_data(sfari_df)
}
}
\keyword{datasets}
