% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{velmeshev_gene_df}
\alias{velmeshev_gene_df}
\title{Velmeshev et al. DEGs}
\format{
## `velmeshev_gene_df`
A data frame with 692 rows (genes) and 14 columns:
\describe{
  \item{Cell type}{Cell type that the DEG was in}
  \item{gene ID}{ENSG gene ID}
  \item{Gene name}{Gene name}
  \item{Gene biotype}{Biotype}
  \item{Fold change}{See original paper for details}
  \item{Sample fold change}{See original paper for details}
  \item{q value}{See original paper for details}
  \item{correlation (bulk mRNA/bulkized nuclear RNA)}{See original paper for details}
  \item{Epilepsy DEG}{boolean}
  \item{gene group}{boolean}
  \item{SFARI gene}{boolean}
  \item{Satterstrom}{boolean}
  \item{Sanders}{boolean}
  \item{cell type-specific expression}{boolean}
}
}
\source{
<https://www.science.org/doi/10.1126/science.aav8130>
}
\usage{
velmeshev_gene_df
}
\description{
From https://www.science.org/doi/10.1126/science.aav8130,
data supplement S4 (i.e., \code{aav8130_data-s4.xls})
}
\examples{
\dontrun{
# How the data was loaded
tib <- readxl::read_xls("aav8130_data-s4.xls",
sheet = "ASD_DEGs")
velmeshev_gene_df <- as.data.frame(tib)
usethis::use_data(velmeshev_gene_df)
}
}
\keyword{datasets}
