% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{compute_mean}
\alias{compute_mean}
\title{Compute mean from natural parameter matrix}
\usage{
compute_mean(nat_mat, family, ...)
}
\arguments{
\item{nat_mat}{matrix}

\item{family}{character}

\item{...}{additional parameters for distribution}
}
\value{
matrix
}
\description{
Compute mean from natural parameter matrix
}
