% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{compute_posterior.default}
\alias{compute_posterior.default}
\title{Compute posterior according to Gamma-Poisson model for matrices and sparse matrices.}
\usage{
\method{compute_posterior}{default}(
  input_obj,
  case_control_variable,
  covariates,
  esvd_res,
  library_size_variable,
  nuisance_vec,
  alpha_max = 50,
  bool_adjust_covariates = F,
  bool_covariates_as_library = T,
  bool_library_includes_interept = T,
  nuisance_lower_quantile = 0.01,
  ...
)
}
\arguments{
\item{input_obj}{Dataset (either \code{matrix} or \code{dgCMatrix}) where the \eqn{n} rows represent cells
and \eqn{p} columns represent genes.
The rows and columns of the matrix should be named.}

\item{case_control_variable}{A string of the column name of \code{covariates} which depicts the case-control
status of each cell. Notably, this should be a binary variable where a \code{1}
is hard-coded to describe case, and a \code{0} to describe control.}

\item{covariates}{\code{matrix} object with \eqn{n} rows with the same rownames as \code{dat} where the columns
represent the different covariates.
Notably, this should contain only numerical columns (i.e., all categorical
variables should have already been split into numerous indicator variables).}

\item{esvd_res}{Output of \code{opt_esvd.default}, notably with elements
\code{x_mat}, \code{y_mat} and \code{z_mat}}

\item{nuisance_vec}{Vector of non-negative numerics of length \code{ncol(input_obj)}, such as
the output of \code{estimate_nuisance.default}.}

\item{alpha_max}{Maximum value of numerator when computing posterior.}

\item{nuisance_lower_quantile}{All the nuisance values that are smaller than this quantile
are set to this quantile.}

\item{...}{Additional parameters.}
}
\value{
A \code{list} of elements \code{posterior_mean_mat}
and \code{posterior_var_mat}
}
\description{
Compute posterior according to Gamma-Poisson model for matrices and sparse matrices.
}
