% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{compute_posterior.eSVD}
\alias{compute_posterior.eSVD}
\title{Compute posterior according to Gamma-Poisson model for eSVD object}
\usage{
\method{compute_posterior}{eSVD}(
  input_obj,
  alpha_max = 50,
  nuisance_lower_quantile = 0.01,
  ...
)
}
\arguments{
\item{input_obj}{\code{eSVD} object outputed from \code{opt_esvd.eSVD}.}

\item{alpha_max}{Maximum value of numerator when computing posterior.}

\item{nuisance_lower_quantile}{All the nuisance values that are smaller than this quantile
are set to this quantile.}

\item{...}{Additional parameters.}
}
\value{
\code{eSVD} object with \code{posterior_mean_mat}
and \code{posterior_var_mat} appended to the list in
\code{input_obj[[input_obj[["latest_Fit"]]]]}.
}
\description{
The posterior is computed based on whatever \code{input_obj$latest_Fit} is set to.
}
