% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_test_statistic.R
\name{compute_test_statistic.default}
\alias{compute_test_statistic.default}
\title{Compute test statistics for matrices}
\usage{
\method{compute_test_statistic}{default}(
  input_obj,
  posterior_var_mat,
  case_individuals,
  control_individuals,
  individual_vec,
  min_cells_per_individual = 3,
  verbose = 0,
  ...
)
}
\arguments{
\item{input_obj}{Posterior mean matrix (a \code{matrix}) where the \eqn{n} rows represent cells
and \eqn{p} columns represent genes.
The rows and columns of the matrix should be named.}

\item{posterior_var_mat}{Posterior variance matrix (a \code{matrix}) where the \eqn{n} rows represent cells
and \eqn{p} columns represent genes.
The rows and columns of the matrix should be the same as those in \code{input_obj}.}

\item{case_individuals}{Vector of strings representing the individuals in \code{metadata[,covariate_individual]}
that are the case individuals.}

\item{control_individuals}{Vector of strings representing the individuals in \code{metadata[,covariate_individual]}
that are the control individuals.}

\item{individual_vec}{Vector of strings of length \eqn{n} (i.e., the number of cells) that denote which cell originates from which individual.}

\item{min_cells_per_individual}{Minimum number of cells an individual must
contribute for the test to be computed. Individuals
with fewer are expected to have been dropped
upstream by \code{eSVD_helper}; reaching here with
one is an error rather than a silently noisy
statistic. Set to \code{0} to disable the check.}

\item{verbose}{Integer.}

\item{...}{Additional parameters.}
}
\value{
A vector of test statistics of length \code{ncol(input_obj)}
}
\description{
Compute test statistics for matrices
}
