% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_test_statistic.R
\name{compute_test_statistic.default}
\alias{compute_test_statistic.default}
\title{Compute test statistics for matrices}
\usage{
\method{compute_test_statistic}{default}(
  input_obj,
  posterior_var_mat,
  case_individuals,
  control_individuals,
  covariate_individual,
  metadata,
  verbose = 0,
  ...
)
}
\arguments{
\item{input_obj}{Posterior mean matrix (a \code{matrix}) where the \eqn{n} rows represent cells
and \eqn{p} columns represent genes.
The rows and columns of the matrix should be named.}

\item{posterior_var_mat}{Posterior variance matrix (a \code{matrix}) where the \eqn{n} rows represent cells
and \eqn{p} columns represent genes.
The rows and columns of the matrix should be the same as those in \code{input_obj}.}

\item{case_individuals}{Vector of strings representing the individuals in \code{metadata[,covariate_individual]}
that are the case individuals.}

\item{control_individuals}{Vector of strings representing the individuals in \code{metadata[,covariate_individual]}
that are the control individuals.}

\item{covariate_individual}{A string of the column name of \code{metadata} which depicts the individual
each cell originates from. Notably, this column in \code{metadata} should be a factor vector.}

\item{metadata}{\code{data frame} object with \eqn{n} rows with the same row names as \code{input_obj$dat}
where the columns represent the different covariates.
Notably, this should can contain categorical variables.}

\item{verbose}{Integer.}

\item{...}{Additional parameters.}
}
\value{
A vector of test statistics of length \code{ncol(input_obj)}
}
\description{
Compute test statistics for matrices
}
