% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_test_statistic.R
\name{compute_test_statistic.eSVD}
\alias{compute_test_statistic.eSVD}
\title{Compute test statistics for eSVD object}
\usage{
\method{compute_test_statistic}{eSVD}(input_obj, verbose = 0, ...)
}
\arguments{
\item{input_obj}{\code{eSVD} object outputed from \code{compute_posterior.eSVD}.}

\item{verbose}{Integer.}

\item{...}{Additional parameters.}
}
\value{
\code{eSVD} object with added element \code{"teststat_vec"}
}
\description{
Compute test statistics for eSVD object
}
