% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{.determine_initial_matrix}
\alias{.determine_initial_matrix}
\title{Initialize the matrix of natural parameters}
\usage{
.determine_initial_matrix(dat, family, k, max_val = NA, tol = 0.001, ...)
}
\arguments{
\item{dat}{dataset where the \code{n} rows represent cells and \code{d} columns represent genes.}

\item{family}{character (\code{"gaussian"}, \code{"exponential"}, \code{"poisson"}, \code{"neg_binom"},
or \code{"curved gaussian"})}

\item{k}{positive integer less than \code{min(c(nrow(dat), ncol(dat)))}}

\item{max_val}{maximum magnitude of the inner product}

\item{tol}{numeric}

\item{...}{extra arguments, such as nuisance parameters for \code{"neg_binom"}
or \code{"curved gaussian"} for \code{family}}
}
\value{
\code{n} by \code{p} matrix
}
\description{
This function first transforms each entry in \code{dat} according to the inverse function that maps
natural parameters to their expectation (according to \code{eSVD:::.mean_transformation}) and then
uses \code{eSVD:::.project_rank_feasibility} to get a rank-\code{k} approximation of this matrix
that lies within the domain of \code{family}
}
