% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance.R
\name{estimate_nuisance.default}
\alias{estimate_nuisance.default}
\title{Estimate nuisance values for matrix or sparse matrices.}
\usage{
\method{estimate_nuisance}{default}(input_obj, mean_mat, library_mat, verbose = 0, ...)
}
\arguments{
\item{input_obj}{Dataset (either \code{matrix} or \code{dgCMatrix}) where the \eqn{n} rows represent cells
and \eqn{p} columns represent genes.
The rows and columns of the matrix should be named.}

\item{mean_mat}{A \code{matrix} of \eqn{n} rows and \eqn{p} columns that represents the
expected value of each entry.}

\item{library_mat}{A \code{matrix} of \eqn{n} rows and \eqn{p} columns that represents the
library size of each entry.}

\item{verbose}{Integer.}

\item{...}{Additional parameters.}
}
\value{
Vector of length \eqn{p}
}
\description{
Estimate nuisance values for matrix or sparse matrices.
}
