% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance.R
\name{estimate_nuisance.eSVD}
\alias{estimate_nuisance.eSVD}
\title{Estimate nuisance values for eSVD objects}
\usage{
\method{estimate_nuisance}{eSVD}(input_obj, verbose = 0, ...)
}
\arguments{
\item{input_obj}{\code{eSVD} object outputed from \code{opt_esvd.eSVD}.
Specifically, the nuisance parameters will be estimated
based on the fit in \code{input_obj[[input_obj[["latest_Fit"]]]]}.}

\item{verbose}{Integer.}

\item{...}{Additional parameters.}
}
\value{
\code{eSVD} object with \code{nuisance_vec} appended to the list in
\code{input_obj[[input_obj[["latest_Fit"]]]]}.
}
\description{
Assumes a Gamma-Poisson model where the mean and variance are proportionally
related.
}
