% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esvd_family.R
\name{esvd_family}
\alias{esvd_family}
\title{Internal Constructor for distribution family}
\usage{
esvd_family(family)
}
\arguments{
\item{family}{string}
}
\value{
an object with conversion functions
}
\description{
Internal Constructor for distribution family
}
