% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_covariates.R
\name{format_covariates}
\alias{format_covariates}
\title{Format covariates}
\usage{
format_covariates(
  dat,
  covariate_df,
  bool_center = FALSE,
  rescale_numeric_variables = NULL,
  variables_enumerate_all = NULL
)
}
\arguments{
\item{dat}{Dataset (either \code{matrix} or \code{dgCMatrix}) where the \eqn{n} rows represent cells
and \eqn{p} columns represent genes.
The rows and columns of the matrix should be named.}

\item{covariate_df}{\code{data.frame} where each row represents a cell, and the
columns are the different categorical or numerical variables that you wish to adjust for}

\item{bool_center}{Boolean if the numerical variables should be centered around zero, default is \code{FALSE}}

\item{rescale_numeric_variables}{A vector of strings denoting the column names in \code{covariate_df} that are numerical and you wish to rescale}

\item{variables_enumerate_all}{If not \code{NULL}, this allows you to control specifically which \code{factor} variables
in \code{covariate_df} you would like to split into indicators. By default, this is \code{NULL}, meaning all the \code{factor} variables are split into indicators}
}
\value{
a \code{matrix} with the same number of rows as \code{dat}
}
\description{
Mainly, this method splits the categorical variables (which should be `factor` variables)
into indicator variables (i.e.,
one-hot encoding), dropping the last level, and then rescales
all the numerical variables (but does not center them),
and computes the \code{"Log_UMI"} (i.e., log total counts) for each cell.
\code{"Log_UMI"} is added as its own column.
}
