% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(
  nat_mat,
  family,
  nuisance_param_vec = NA,
  library_size_vec = 1,
  tol = 0.001
)
}
\arguments{
\item{nat_mat}{An \eqn{n\times p} matrix of natural parameters, where
\eqn{n} rows represent cells and \eqn{p} columns represent genes.}

\item{family}{A character string, one of \code{"gaussian"}, \code{"exponential"},
\code{"poisson"}, \code{"neg_binom"}, \code{"neg_binom2"},
\code{"curved_gaussian"}, and \code{"bernoulli"}.}

\item{nuisance_param_vec}{Either \code{NA} or a single numeric or a length-\eqn{p}
vector of numerics representing nuisance parameters.
It is required if \code{family} is \code{"gaussian"},
\code{"curved_gaussian"}, \code{"neg_binom"} or
\code{"neg_binom2"}, and ignored otherwise.}

\item{library_size_vec}{A single positive numeric (recycled to every cell; default \code{1})
or a length-\eqn{n} vector of positive numerics.}

\item{tol}{Small positive value to determine the smallest possible value in the output
matrix, applied only when \code{family} is \code{"gaussian"} or
\code{"curved_gaussian"}.}
}
\value{
The generated data matrix, of the same dimension as \code{nat_mat}
}
\description{
Generate data
}
\examples{
set.seed(10)
x_mat <- matrix(stats::rnorm(20 * 2), nrow = 20, ncol = 2)
y_mat <- matrix(stats::rnorm(10 * 2), nrow = 10, ncol = 2)
nat_mat <- tcrossprod(x_mat, y_mat)
dat <- generate_data(nat_mat = nat_mat, family = "poisson")
dim(dat)
}
