% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(
  nat_mat,
  family,
  nuisance_param_vec = NA,
  library_size_vec = NA,
  tol = 0.001
)
}
\arguments{
\item{nat_mat}{An \eqn{n\times p}{n x p} matrix of natural parameters, where
\eqn{n} rows represent cells and \eqn{p} columns represent genes.}

\item{family}{A character string, one of \code{"gaussian"}, \code{"exponential"},
\code{"poisson"}, \code{"neg_binom"}, \code{"curved_gaussian"},
and \code{"bernoulli"}.}

\item{nuisance_param_vec}{Either \code{NA} or a single numeric or a length-\eqn{p}
vector of numerics representing nuisance parameters
(for \code{family = "neg_binom"} and
\code{family = "curved_gausian"}).
It is only required if
\code{family \%in\% c("neg_binom", "curved_gaussian")}.}

\item{library_size_vec}{Either \code{NA} or a length-\eqn{n} vector of numerics}

\item{tol}{Small positive value to determine the smallest possible value in the output
matrix, useful for only \code{family = "curved_gaussian"}.}
}
\value{
The generated data matrix
}
\description{
Generate data
}
