% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(
  nat_mat,
  family,
  nuisance_param_vec = NA,
  library_size_vec = NA,
  tol = 0.001
)
}
\arguments{
\item{nat_mat}{matrix of natural parameters where the \code{n} rows represent cells and \code{d} columns represent genes}

\item{family}{character (\code{"gaussian"}, \code{"exponential"}, \code{"poisson"}, \code{"neg_binom"},
or \code{"curved_gaussian"})}

\item{nuisance_param_vec}{either \code{NA} or a single numeric or a length-\code{p} 
vector of numerics representing nuisance parameters (for \code{family = "neg_binom"} or 
\code{family = "curved_gausian"}). It is only required if \code{family \%in\% c("neg_binom", "curved_gaussian")}}

\item{library_size_vec}{either \code{NA} or a length-\code{n} vector of numerics}

\item{tol}{small positive value to determine the smallest possible value in the output matrix, 
useful for only \code{family="curved_gaussian"}}
}
\value{
matrix
}
\description{
Generate data
}
