% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_null.R
\name{generate_null}
\alias{generate_null}
\title{Generate null data}
\usage{
generate_null(cell_per_person = 100, num_genes = 1000, num_individuals = 20)
}
\arguments{
\item{cell_per_person}{The number of cells per individual}

\item{num_genes}{The number of genes in the simulation}

\item{num_individuals}{The number of individuals}
}
\value{
a list with \code{covariates} (a `matrix` with `n` cells and
4 columns, named `"Intercept"`, `"Log_UMI"`, `"Sex"`, and `"Age"`,
which will be used in `eSVD2::initialize_esvd`),
\code{df} (a `data.frame` with `num_individuals` rows
and 5 columns, named named `"Intercept"`, `"Log_UMI"`, `"Sex"`, `"Age"`, and `"Individual"`),
\code{metadata_individual} (a `matrix` with `n` rows and 1 column called `"Individual"` for
which cell originates from which individual)
\code{nat_mat} (a `n` by `p` `matrix` denoting the natural parameter for each cell's gene expression), and
\code{obs_mat} (a `n` by `p` `dgCMatrix` denoting the observed count for each cell's gene expression),
}
\description{
This function doesn't allow of much flexbility, and is meant primarily for simple null simulations.
Here, `n` (the number of cells) is equal to `cell_per_person*num_individuals`.
}
