% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initialize_esvd}
\alias{initialize_esvd}
\title{Initialize eSVD}
\usage{
initialize_esvd(
  dat,
  k,
  family,
  nuisance_param_vec = NA,
  library_size_vec = NA,
  config = initalization_default()
)
}
\arguments{
\item{dat}{dataset where the \code{n} rows represent cells and \code{d} columns represent genes}

\item{k}{positive integer less than \code{min(c(nrow(dat), ncol(dat)))}}

\item{family}{character (\code{"gaussian"}, \code{"exponential"}, \code{"poisson"}, \code{"neg_binom"},
or \code{"curved_gaussian"})}

\item{nuisance_param_vec}{either \code{NA} or a single numeric or a length-\code{p} 
vector of numerics representing nuisance parameters (for \code{family = "neg_binom"} or 
\code{family = "curved_gausian"})}

\item{library_size_vec}{either \code{NA} or a length-\code{n} vector of numerics}

\item{config}{additional parameters for the initialization, whose defaults can be
set with \code{eSVD2::initialization_default()}}
}
\value{
a list with elements \code{x_mat} and \code{y_mat}, representing the two
latent matrices
}
\description{
Initialize eSVD
}
