% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initialize_esvd}
\alias{initialize_esvd}
\title{Initialize eSVD}
\usage{
initialize_esvd(
  dat,
  k,
  family,
  covariates = NULL,
  offset_vec = rep(0, nrow(dat)),
  verbose = 0
)
}
\arguments{
\item{dat}{dataset where the \eqn{n} rows represent cells and \eqn{p} columns represent genes}

\item{k}{positive integer less than \code{min(c(nrow(dat), ncol(dat)))}}

\item{family}{character (\code{"gaussian"}, \code{"exponential"}, \code{"poisson"}, \code{"neg_binom"},
or \code{"curved_gaussian"})}

\item{covariates}{an \eqn{n \times d}{n × d} matrix representing the additional \eqn{d} covariates,
or \code{NULL} if no covariate is given}

\item{offset_vec}{a vector of length-\eqn{n} that represents a constant amount added to each row of the
natural parameter matrix}

\item{verbose}{non-negative integer specifying level of printouts}
}
\value{
a list with elements \code{x_mat} and \code{y_mat} (and others), representing the two
latent matrices
}
\description{
Initialize eSVD
}
