% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multtest.R
\name{multtest}
\alias{multtest}
\title{Perform multiple-testing adjustment using Efron's empirical null}
\usage{
multtest(teststat_vec, observed_quantile = c(0.05, 0.95))
}
\arguments{
\item{teststat_vec}{a vector of Gaussian-like test statistics}

\item{observed_quantile}{The quantiles of \code{teststat_vec} used to estimate the null distribution}
}
\value{
a list with \code{fdr_vec}, \code{logpvalue_vec},
\code{method}, \code{null_mean}, \code{null_sd}, and \code{pvalue_vec}
}
\description{
Perform multiple-testing adjustment using Efron's empirical null
}
