% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{opt_esvd.default}
\alias{opt_esvd.default}
\title{Optimize eSVD for matrices or sparse matrices.}
\usage{
\method{opt_esvd}{default}(
  input_obj,
  x_init,
  y_init,
  z_init = NULL,
  covariates = NULL,
  family = "poisson",
  l2pen = 0.1,
  library_multipler = rep(1, nrow(input_obj)),
  max_iter = 100,
  nuisance_vec = rep(NA, ncol(input_obj)),
  offset_variables = NULL,
  tol = 1e-06,
  verbose = 0,
  ...
)
}
\arguments{
\item{input_obj}{Dataset (either \code{matrix} or \code{dgCMatrix}) where the \eqn{n} rows represent cells
and \eqn{p} columns represent genes.
The rows and columns of the matrix should be named.}

\item{x_init}{Initial matrix of the cells' latent vectors that is \eqn{n} rows and \eqn{k}
columns. The row names should be the same as \code{input_obj}.}

\item{y_init}{Initial matrix of the genes' latent vectors that is \eqn{p} rows and \eqn{k}
columns. The row names should be the same as the column names of \code{input_obj}.}

\item{z_init}{Initial matrix of the genes' coefficient vectors that is \eqn{p} rows and \code{ncol(covariates)}
columns. The row names should be the same as the column names of \code{input_obj},
and the column names should be the same as \code{covariates}.}

\item{covariates}{\code{matrix} object with \eqn{n} rows with the same rownames as \code{input_obj} where the columns
represent the different covariates.
Notably, this should contain only numerical columns (i.e., all categorical
variables should have already been split into numerous indicator variables).}

\item{family}{String among \code{"gaussian"}, \code{"curved_gaussian"},
\code{"exponential"}, \code{"poisson"}, \code{"neg_binom"},
\code{"neg_binom2"}, or \code{"bernoulli"}. Notably, with exception of
\code{"neg_binom2"}, all the other families are parameterized such that
eSVD is fitting the dot product to be the canonical parameter of these
expoential-family distributions. For \code{"neg_binom2"}, the dot
product is the log-mean of the distribution (i.e., similar to the canonical
parameterization of the Poisson family).}

\item{l2pen}{Small positive number for the amount of penalization for both the cells'
and the genes' latent vectors as well as the coefficients.}

\item{library_multipler}{Vector of positive numerics of length \eqn{n}. It is the multiplier
such that the variance of cell \code{i}'s entries is the mean of
cell \code{i}'s entries times the square-root of cell \code{i}'s
value in \code{library_multipler} (entry-wise). This is used as
an alternative interpretation of how library-size affects a cell's
gene expression (instead of using the library size as a covariate to be
regressed out).}

\item{max_iter}{Positive integer for number of iterations.}

\item{nuisance_vec}{Vector of non-negative numerics (or \code{NA}'s) of length \eqn{p},
representing each gene's nuisance parameter when using an exponential-family
distribution that requires one.
It is used only when \code{family} is  \code{"curved_gaussian"} or
\code{"neg_binom"} or \code{"neg_binom2"}.}

\item{offset_variables}{A vector of strings depicting which column names in \code{input_obj$covariate}
be treated as an offset during the optimization (i.e., their coefficients will not change
throughout the optimization).}

\item{tol}{Small positive number to differentiate between zero and non-zero.}

\item{verbose}{Integer.}

\item{...}{Additional parameters.}
}
\value{
a \code{list} with elements \code{x_mat}, \code{y_mat},
\code{z_mat}, \code{library_multiplier}, \code{loss}, \code{nuisance_vec}
and \code{param}.
}
\description{
Optimize eSVD for matrices or sparse matrices.
}
