% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{opt_esvd.eSVD}
\alias{opt_esvd.eSVD}
\title{Optimize eSVD for eSVD objects}
\usage{
\method{opt_esvd}{eSVD}(
  input_obj,
  fit_name = "fit_First",
  fit_previous = "fit_Init",
  l2pen = 0.1,
  max_iter = 100,
  offset_variables = NULL,
  tol = 1e-06,
  verbose = 0,
  ...
)
}
\arguments{
\item{input_obj}{\code{eSVD} object outputed from \code{apply_initial_threshold}.}

\item{fit_name}{String for the name of that will become the current fit when
storing the results in \code{input_obj}.}

\item{fit_previous}{String for the name of the previous fit that this function will
grab the initialization values from.}

\item{l2pen}{Small positive number for the amount of penalization for both the cells'
and the genes' latent vectors as well as the coefficients.}

\item{max_iter}{Positive integer for number of iterations.}

\item{offset_variables}{A vector of strings depicting which column names in \code{input_obj$covariate}
be treated as an offset during the optimization (i.e., their coefficients will not change
throughout the optimization).}

\item{tol}{Small positive number to differentiate between zero and non-zero.}

\item{verbose}{Integer.}

\item{...}{Additional parameters.}
}
\value{
\code{eSVD} object with added elements with name to whatever
\code{fit_name} was set to.
}
\description{
Optimize eSVD for eSVD objects
}
