% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_helper.R
\name{opt_x}
\alias{opt_x}
\title{Optimize X given C, Y and Z}
\usage{
opt_x(
  XC_init,
  YZ,
  k,
  loader,
  family,
  s,
  gamma,
  l2penx,
  verbose = 0,
  inplace = FALSE,
  ...
)
}
\arguments{
\item{XC_init}{Initial value for the `[X C]` matrix, X [n x k], C [n x r]}

\item{YZ}{The `[Y Z]` matrix, Y [p x k], Z [p x r]}

\item{k}{Number of columns in X and Y}

\item{loader}{The data loader, typically returned by data_loader()}

\item{family}{A family object, typically returned by esvd_family()}

\item{s}{The library size vector, [n x 1]}

\item{gamma}{The nuisance parameter vector, [p x 1]}

\item{l2penx}{The l2 penalty parameter for X, a scalar}

\item{verbose}{Verbosity parameter}

\item{inplace}{Whether the input XC_init will be modified and returned}

\item{...}{Additional parameters}
}
\description{
Optimize X given C, Y and Z
}
