% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loader.R
\name{print.esvd_data_loader}
\alias{print.esvd_data_loader}
\title{Internal data loader function}
\usage{
\method{print}{esvd_data_loader}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{additional arguments}
}
\value{
description
}
\description{
Internal data loader function
}
