% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparameterization.R
\name{reparameterization_esvd_covariates}
\alias{reparameterization_esvd_covariates}
\title{Reparameterize eSVD object}
\usage{
reparameterization_esvd_covariates(
  input_obj,
  fit_name,
  omitted_variables = NULL,
  verbose = 0
)
}
\arguments{
\item{input_obj}{\code{eSVD} object, after either \code{initialize_esvd()} or
\code{opt_esvd()}}

\item{fit_name}{The name of the fit in \code{input_obj} that you wish to reparameterize. This should be in \code{names(input_obj)}}

\item{omitted_variables}{Either \code{NULL} (the default) or variables in \code{input_obj$covariates} that should not be reparameterized}

\item{verbose}{Integer.}
}
\value{
\code{eSVD} object after adjusting the fit in \code{fit_name}.
}
\description{
Reparameterize eSVD object
}
