% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_results.R
\name{report_results}
\alias{report_results}
\title{Report the DE results from eSVD}
\usage{
report_results(input_obj)
}
\arguments{
\item{input_obj}{\code{eSVD} object outputed from \code{compute_pvalue}.}
}
\value{
a data frame
}
\description{
Report the DE results from eSVD
}
